requirements.txt
flask
flask-cors
requests
python-dotenv
youtube-transcript-api
